% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splinefit.R
\name{splinefit}
\alias{splinefit}
\title{Automatic spline-based model selection}
\usage{
splinefit(
  x,
  y,
  nlimknot,
  lambda,
  delta,
  degree,
  bspline = TRUE,
  b.knots = NULL
)
}
\arguments{
\item{x}{Vector of function arguments.}

\item{y}{Vector of function values.}

\item{nlimknot}{Maximum number of internal knots.}

\item{lambda}{Tunning parameter of the penalized least squares estimation process.}

\item{delta}{Minimal distance value between two consecutive knots.}

\item{degree}{Degree of the spline basis.}

\item{bspline}{External knots vector. Default is the (min(x),max(x)).}

\item{b.knots}{Logical value. If TRUE, B-spline is fitted. Else, natural splines.}
}
\value{
A list with selected knots and fitted model.
}
\description{
Provides optimal number of internal knots and their locations in spline-based models.
}
