\name{splineSelection-package}
\alias{splineSelection-package}
\alias{splineSelection}
\docType{package}
\title{\packageTitle{splineSelection}}
\description{\packageDescription{splineSelection}}
\section{Package Content}{\packageIndices{splineSelection}}
\author{\packageAuthor{splineSelection}}
\section{Maintainer}{\packageMaintainer{splineSelection}}
\keyword{package}
